API

Types

MulticomplexNumbers.Multicomplex — Type

Multicomplex{T,N,C}(value)

Defines a multicomplex number ℂ_N, over base field T, having C=2^N components specified in static vector value.

Input: value::SVector{C, T})

Constants

MulticomplexNumbers.im1 — Constant

im1, im2, im3, im4, im5, im6

The imaginary units, up to dimension 6.

Functions

MulticomplexNumbers.order — Function

Order of the multicomplex number is the number of imaginary units

MulticomplexNumbers.flat — Function

Get components as a vector

MulticomplexNumbers.component — Function

Utility function to extract a real-valued component from a multicomplex number

MulticomplexNumbers.realest — Function

Extract the 'most real' component

MulticomplexNumbers.matrep — Function

matrix representation of a multicomplex number

MulticomplexNumbers.ascomplex — Function

ascomplex(A::AbstractArray{M}, [unit])

Returns a view of the multicomplex input array A as an array of complex numbers, mapping i(unit) -> im. If a unit is not supplied, defaults to the highest order.

If A has size (m, n, ...) with multicomplex numbers of dimension N, then the output will have size (2^(N-1), m, n, ...).

Index

MulticomplexNumbers.im1

MulticomplexNumbers.Multicomplex

MulticomplexNumbers.ascomplex

MulticomplexNumbers.component

MulticomplexNumbers.flat

MulticomplexNumbers.matrep

MulticomplexNumbers.order

MulticomplexNumbers.realest