API

Module

Missing docstring.

Missing docstring for MulticomplexNumbers. Check Documenter's build log for details.

Types

Missing docstring.

Missing docstring for MulticomplexNumber. Check Documenter's build log for details.

Functions and methods

MulticomplexNumbers.Multicomplex — Method

Multicomplex{N}(value)

Defines a multicomplex number ℂ_N, with base field and components defined in value.

Input: value::SVector{C, T})

MulticomplexNumbers.Multicomplex — Method

Multicomplex{N}(value)

Defines a multicomplex number ℂ_N, with base field and components defined in value.

Input: value::SVector{C, T})

Build a higher order multicomplex number from two multicomplex 'real' and 'imaginary' components

Missing docstring.

Missing docstring for Multicomplex(x::Real, y::Real). Check Documenter's build log for details.

MulticomplexNumbers.Multicomplex — Method

Multicomplex{N}(value)

Defines a multicomplex number ℂ_N, with base field and components defined in value.

Input: value::SVector{C, T})

Build a higher order multicomplex number from two multicomplex 'real' and 'imaginary' components

MulticomplexNumbers.Multicomplex — Method

Multicomplex(x,y,u,v) = x + i¹y + i²u + i¹i²v

Missing docstring.

Missing docstring for Multicomplex(a::Complex, b::Complex). Check Documenter's build log for details.

MulticomplexNumbers.Multicomplex — Method

Build a higher order multicomplex number from two multicomplex 'real' and 'imaginary' components

MulticomplexNumbers.component — Method

Utility function to extract a real-valued component from a multicomplex number

MulticomplexNumbers.realest — Method

Extract the 'most real' component

Missing docstring.

Missing docstring for Base.real(m::Multicomplex{T,N,C}) where {T,N,C}. Check Documenter's build log for details.

MulticomplexNumbers.matrep — Method

matrix representation of a multicomplex number

MulticomplexNumbers.ascomplex — Method

ascomplex(A::AbstractArray{M})

Returns a view of the multicomplex input array A as an array of complex numbers, mapping i⁽¹⁾ -> im

If A has size (m, n, ...) with multicomplex numbers of dimension N, then the output will have size (2^(N-1), m, n, ...).

Base.:* — Method

*(a,b)

Multicomplex multiplication via the matrix representation.

Missing docstring.

Missing docstring for conj(m::Multicomplex) = Multicomplex(real(m),-imag(m)). Check Documenter's build log for details.

Index

MulticomplexNumbers.Multicomplex

MulticomplexNumbers.Multicomplex

MulticomplexNumbers.Multicomplex

MulticomplexNumbers.Multicomplex

MulticomplexNumbers.Multicomplex

Base.:*

MulticomplexNumbers.ascomplex

MulticomplexNumbers.component

MulticomplexNumbers.matrep

MulticomplexNumbers.realest